// backend/src/films/dto/films.dto.ts
//TODO описать DTO для запросов к /films

export class FilmDto {
  id: string;
  rating: number;
  director: string;
  tags: string[];
  title: string;
  about: string;
  description: string;
  image: string;
  cover: string;
}

export class FilmResponseDto {
  total: number;
  items: FilmDto[];
}

export class ScheduleDto {
  id: string;
  daytime: string;
  hall: string;
  rows: number;
  seats: number;
  price: number;
  taken: string[];
}

export class ScheduleResponseDto {
  total: number;
  items: ScheduleDto[];
}